fix(sonar): clear new-code quality gate on fbs-gen merge
Two gate conditions failed after the in-house FlatBuffers toolchain landed:
new_reliability_rating (6 BUGs) and new_duplicated_lines_density.

Reliability — 5× S2184 (int offset arithmetic widened to a long index, a
real overflow guard) and 1× S6218:
- FbsBuilder vtable dedup/patch: widen the segment-index operands to long
  (the soffset *value* at the patch site stays int by design).
- FlatSegmentDecoder / InspectorTree peekFlatRoot: read the trailing length
  field at `segLen - 4L`; InspectorTree's fbStart becomes long to match its
  bounds-checked twin in FlatSegmentDecoder.
- TypeRegistry.StructLayout: @SuppressWarnings("java:S6218"); codegen-internal
  carrier, fieldOffsets never compared (same precedent as UnpackSchedule).

Duplication — exclude the two build-only code generators (fbs-gen, proto-gen)
from CPD via sonar.cpd.exclusions. Their lexers share tokenizer shape by
design; smell and bug analysis on them is kept.

+++ inspector/src/main/java/io/github/dfa1/vortex/inspect/InspectorTree.java
@@ -255,8 +255,8 @@ public record InspectorTree(
 
     private static Peek peekFlatRoot(MemorySegment seg, List<String> arraySpecs) {
         int segLen = (int) seg.byteSize();
-        int fbLen = seg.get(LE_INT, segLen - 4);
-        int fbStart = segLen - 4 - fbLen;
+        int fbLen = seg.get(LE_INT, segLen - 4L);
+        long fbStart = segLen - 4L - fbLen;
         FbsArray fbArray = FbsArray.getRootAsFbsArray(seg.asSlice(fbStart, fbLen));
         FbsArrayNode root = fbArray.root();
         if (root == null) {
